Add more details to revision tooltip
