+++ content/less/recipe-details-page.less
@@ -41,3 +41,7 @@
     }
   }
 }
+
+.history-timeline-tooltip {
+  margin-top: -5px;
+}
